Floats 2

+++ src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
     let my_number = 9.; // f64
     let other_number = 9; // i32
-    println!("{}", my_number as i32 + other_number);
+    println!("{}", my_number + other_number as f64);
 }
